Startup Sound.m4a [Unofficial]
An unofficial fan-made Windows 10 startup sound, in M4A format. This is not the official startup sound for Windows 10. All versions of Windows from Windows Vista onward use the same startup and shutdown sound, but it is disabled by default on Windows 8 and onward. I am not sure about Server 2012/2016/2019 though.